back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
pymongo==4.6.0
email-validator==2.1.0
python-multipart==0.0.6 